+++ src/main/scala/net/rfc1149/rxtelegram/model/GroupChat.scala
@@ -2,7 +2,7 @@ package net.rfc1149.rxtelegram.model
 
 import play.api.libs.json.{Reads, Json, Format}
 
-case class GroupChat (id: Int, title: String) extends Conversation
+case class GroupChat (id: Long, title: String) extends Conversation
 
 object GroupChat {
   implicit val groupChatFormat: Format[GroupChat] = Json.format[GroupChat]
