+++ src/main/scala/net/rfc1149/rxtelegram/model/PhotoSize.scala
@@ -2,7 +2,7 @@ package net.rfc1149.rxtelegram.model
 
 import play.api.libs.json.{Json, Format}
 
-case class PhotoSize(file_id: String, width: Int, height: Int, file_size: Option[Int])
+case class PhotoSize(file_id: String, width: Long, height: Long, file_size: Option[Long])
 
 object PhotoSize {
   implicit val photoSizeFormat: Format[PhotoSize] = Json.format[PhotoSize]
