+++ src/main/scala/net/rfc1149/rxtelegram/model/UserProfilePhotos.scala
@@ -2,7 +2,7 @@ package net.rfc1149.rxtelegram.model
 
 import play.api.libs.json.{Json, Format}
 
-case class UserProfilePhotos(total_count: Int, photos: Array[Array[PhotoSize]])
+case class UserProfilePhotos(total_count: Long, photos: Array[Array[PhotoSize]])
 
 object UserProfilePhotos {
   implicit val userProfilePhotosFormat: Format[UserProfilePhotos] = Json.format[UserProfilePhotos]
